pyproject.toml
[tool.poetry]
name = "inference-api"
version = "0.1.0"
description = "Skin lesion inference API"
authors = ["Rob Pickerill <rob@spotplotter.com>"]
homepage = "https://spotplotter.com"
repository = "https://github.com/spot-plot/inference-api"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
tensorflow = "^2.18.0"
numpy = "^2.0.0"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
pillow = "^11.0.0"
python-multipart = "^0.0.20"

[tool.poetry.dev-dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
isort = "^5.13.2"

[tool.poetry.group.dev.dependencies]
uvicorn = "^0.34.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
